Fix int return values
- Go json unmarshalls numbers as float64 even for ints, see https://stackoverflow.com/questions/55436628/json-decoded-value-is-treated-as-float64-instead-of-int/55436758
- manually cast them to int after getting float values to overcome this

+++ configurt.go
@@ -105,7 +105,7 @@ func (c *Client) GetAsFloat(configKey string) float64 {
 // GetAsFloat returns config value as int
 func (c *Client) GetAsInt(configKey string) int {
 	val := c.Get(configKey)
-	return val.(int)
+	return int(val.(float64))
 }
 
 // GetAsStringArray returns config value as string array
@@ -138,7 +138,7 @@ func (c *Client) GetAsIntArray(configKey string) []int {
 	array := val.([]interface{})
 	intArray := []int{}
 	for _, v := range array {
-		intv := v.(int)
+		intv := int(v.(float64))
 		intArray = append(intArray, intv)
 	}
 	return intArray
